move the logic for GET/PUT on quotas to type Processor
I want to reuse this for the upcoming implementation of the LIQUID protocol.
